Fix BF16 validation loss overflow for continue-training
The validation loss was computed inside torch.amp.autocast (BF16), but
pretrained models produce confident logits that overflow exp() in the
cross-entropy softmax under BF16's limited precision. This produced
val_loss values in the hundreds or thousands while per-class losses
(computed in FP32 with clamping) were reasonable.

Fix: compute validation loss in FP32 by casting outputs to float before
the criterion. The forward pass still runs in BF16 for speed; only the
loss computation is promoted to FP32 for numerical stability.

Bump version to 0.4.1

+++ python_server/dlclassifier_server/services/training_service.py
@@ -1834,7 +1834,10 @@ class TrainingService:
                     if use_mixed_precision:
                         with torch.amp.autocast("cuda", dtype=amp_dtype):
                             outputs = model(images)
-                            loss = criterion(outputs, masks)
+                        # Compute val loss in FP32 to avoid BF16 overflow.
+                        # BF16 logits from confident pretrained models can
+                        # overflow exp() in cross-entropy, producing inf loss.
+                        loss = criterion(outputs.float(), masks)
                     else:
                         outputs = model(images)
                         loss = criterion(outputs, masks)
